+++ Makefile
@@ -4,7 +4,7 @@ INCDIR := $(TOPDIR)
 
 CC = $(CROSS_COMPILE)gcc
 STRIP = $(CROSS_COMPILE)strip
-CFLAGS += -Wall -Os -I$(INCDIR) -O2 -D_FILE_OFFSET_BITS=64
+CFLAGS += -Wall -I$(INCDIR) -O2 -D_FILE_OFFSET_BITS=64
 #CFLAGS += -Wall -Os -I$(INCDIR) -g -D_FILE_OFFSET_BITS=64
 
 export CC STRIP CFLAGS TOPDIR INCDIR
