+++ engine/esma_reactor.h
@@ -16,14 +16,6 @@ api_declaration(reactor) {
 	 int (*del)(int fd, struct esma_channel *ch);
 	 int (*mod)(int fd, struct esma_channel *ch, u32 events);
 
-	 /* signal section */
-	 int (*new_sig)(int sig);	/* returned new sigfd or -1 */
-
-	 /* timers section */
-	 int    (*new_timerfd)(void);	/* returned new timerfd or -1 */
-	 int    (*arm_timerfd)(int timerfd, int time_interval, int type);
-	 int (*disarm_timerfd)(int timerfd);
-
 	void (*wait)();
 };
